# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reporting/sales-chart": {
      "get": {
        "operationId": "ReportingController_salesChart",
        "parameters": [
          {
            "name": "granularity",
            "required": false,
            "in": "query",
            "schema": {
              "example": "q6",
              "type": "string",
              "enum": [
                "day",
                "month",
                "q3",
                "q6",
                "year",
                "flight"
              ]
            }
          },
          {
            "name": "periodStart",
            "required": false,
            "in": "query",
            "description": "Required for granularity=month — ISO date of the 1st of that month (UTC)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Required for granularity=day — ISO date (UTC)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flightNo",
            "required": false,
            "in": "query",
            "description": "Required for granularity=flight",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Channel-split sales bars, shared identically by all 6 panels dashboards",
        "tags": [
          "reporting"
        ]
      }
    },
    "/reporting/kpis": {
      "get": {
        "operationId": "ReportingController_kpis",
        "parameters": [
          {
            "name": "granularity",
            "required": false,
            "in": "query",
            "schema": {
              "example": "q6",
              "type": "string",
              "enum": [
                "day",
                "month",
                "q3",
                "q6",
                "year",
                "flight"
              ]
            }
          },
          {
            "name": "periodStart",
            "required": false,
            "in": "query",
            "description": "Required for granularity=month",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Required for granularity=day",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flightNo",
            "required": false,
            "in": "query",
            "description": "Required for granularity=flight",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "periodKey",
            "required": false,
            "in": "query",
            "description": "A periodKey returned by /reporting/sales-chart — re-scopes KPIs to that bucket",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Revenue/profit/cost KPI boxes — re-scopes to periodKey when provided",
        "tags": [
          "reporting"
        ]
      }
    },
    "/reporting/completed-flights-summary": {
      "get": {
        "operationId": "ReportingController_completedFlightsSummary",
        "parameters": [
          {
            "name": "granularity",
            "required": false,
            "in": "query",
            "schema": {
              "example": "q6",
              "type": "string",
              "enum": [
                "day",
                "month",
                "q3",
                "q6",
                "year",
                "flight"
              ]
            }
          },
          {
            "name": "periodStart",
            "required": false,
            "in": "query",
            "description": "Required for granularity=month",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Required for granularity=day",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flightNo",
            "required": false,
            "in": "query",
            "description": "Required for granularity=flight",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "periodKey",
            "required": false,
            "in": "query",
            "description": "A periodKey returned by /reporting/sales-chart — re-scopes KPIs to that bucket",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Completed-flight seat stats, synced to the same period as the sales chart",
        "tags": [
          "reporting"
        ]
      }
    },
...
    "/reporting/revenue-mix": {
      "get": {
        "operationId": "ReportingController_revenueMix",
        "parameters": [
          {
            "name": "granularity",
            "required": false,
            "in": "query",
            "schema": {
              "example": "q6",
              "type": "string",
              "enum": [
                "day",
                "month",
                "q3",
                "q6",
                "year",
                "flight"
              ]
            }
          },
          {
            "name": "periodStart",
            "required": false,
            "in": "query",
            "description": "Required for granularity=month",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "Required for granularity=day",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flightNo",
            "required": false,
            "in": "query",
            "description": "Required for granularity=flight",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "periodKey",
            "required": false,
            "in": "query",
            "description": "A periodKey returned by /reporting/sales-chart — re-scopes KPIs to that bucket",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "ترکیب درآمد بر اساس کانال فروش",
        "tags": [
          "reporting"
        ]
      }
    },
...
  },
  "info": {
    "title": "blujet API",
    "description": "blujet airline platform — internal API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}